Spinlocks: de-duplicate code across architectures
With this change, spinlock-related definitions are common across
architectures. Only the implementation of atomic operations and
fences (in functions such as fetch_and_add()) is
architecture-specific.
