+++ src/aarch64/kernel_machine.h
@@ -191,7 +191,6 @@ static inline void wait_for_interrupt(void)
 }
 
 /* locking constructs */
-#include <lock.h>
 #include <mutex.h>
 
 /* device mmio region access */
