+++ src/kernel/lockstats_struct.h
@@ -1,3 +1,6 @@
+#define LOCK_TYPE_SPIN 0
+#define LOCK_TYPE_MUTEX 1
+
 typedef struct lockstats_lock {
     int type;
     u64 acq_time;
