+++ src/riscv64/kernel_machine.h
@@ -129,7 +129,6 @@ static inline void wait_for_interrupt(void)
 }
 
 /* locking constructs */
-#include <lock.h>
 #include <mutex.h>
 
 /* device mmio region access */
