+++ src/riscv64/machine.h
@@ -130,10 +130,6 @@ static inline u8 tagof(void *v)
 
 #endif
 
-typedef struct spinlock {
-    word w;
-} *spinlock;
-
 /* returns -1 if x == 0, caller must check */
 static inline __attribute__((always_inline)) u64 msb(u64 x)
 {
