+++ src/runtime/runtime.h
@@ -60,6 +60,8 @@ void print_frame_trace_from_here();
 
 #define build_assert(x) _Static_assert((x), "build assertion failure")
 
+#include <lock.h>
+
 void runtime_memcpy(void *a, const void *b, bytes len);
 
 void runtime_memset(u8 *a, u8 b, bytes len);
