+++ src/x86_64/kernel_machine.h
@@ -164,7 +164,6 @@ void install_gdt64_and_tss(void *tss_desc, void *tss, void *gdt, void *gdt_point
 
 #ifdef KERNEL
 /* locking constructs */
-#include <lock.h>
 #include <mutex.h>
 #endif
