+++ src/x86_64/machine.h
@@ -41,22 +41,6 @@ static inline u8 tagof(void *v)
 }
 
 #endif
-
-#ifdef LOCK_STATS
-#include <lockstats_struct.h>
-#endif
-
-typedef struct spinlock {
-    word w;
-#ifdef LOCK_STATS
-    struct lockstats_lock s;
-#endif
-} *spinlock;
-
-typedef struct rw_spinlock {
-    struct spinlock l;
-    u64 readers;
-} *rw_spinlock;
 #endif
 
 static inline __attribute__((always_inline)) void compiler_barrier(void)
